frontend_service/requirements.txt
Django==5.2.3
djangorestframework
channels
requests
djangorestframework-simplejwt
psycopg2-binary